HW6_CSCI571/requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
python-dateutil==2.8.2
requests==2.27.1